requirements/development.txt
-r common.txt

django-debug-toolbar==3.6.0
django-extensions